Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/css/css-transforms/animation/transform-perspective-composition-expected.txt


FAIL Compositing: property <transform> underlying [perspective(10px)] from add [perspective(10px)] to add [perspective(50px)] at (-0.5) should be [perspective(4.12px)] assert_equals: expected "matrix3d ( 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 0 , 1 , - 0.24 , 0 , 0 , 0 , 1 ) " but got "matrix3d ( 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 0 , 1 , - 0.14 , 0 , 0 , 0 , 1 ) "
FAIL Compositing: property <transform> underlying [perspective(10px)] from add [perspective(10px)] to add [perspective(50px)] at (0) should be [perspective(5px)] assert_equals: expected "matrix3d ( 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 0 , 1 , - 0.2 , 0 , 0 , 0 , 1 ) " but got "matrix3d ( 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 0 , 1 , - 0.1 , 0 , 0 , 0 , 1 ) "
FAIL Compositing: property <transform> underlying [perspective(10px)] from add [perspective(10px)] to add [perspective(50px)] at (0.25) should be [perspective(5.45px)] assert_equals: expected "matrix3d ( 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 0 , 1 , - 0.18 , 0 , 0 , 0 , 1 ) " but got "matrix3d ( 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 0 , 1 , - 0.08 , 0 , 0 , 0 , 1 ) "
FAIL Compositing: property <transform> underlying [perspective(10px)] from add [perspective(10px)] to add [perspective(50px)] at (0.5) should be [perspective(6.15px)] assert_equals: expected "matrix3d ( 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 0 , 1 , - 0.16 , 0 , 0 , 0 , 1 ) " but got "matrix3d ( 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 0 , 1 , - 0.06 , 0 , 0 , 0 , 1 ) "
FAIL Compositing: property <transform> underlying [perspective(10px)] from add [perspective(10px)] to add [perspective(50px)] at (0.75) should be [perspective(7.06px)] assert_equals: expected "matrix3d ( 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 0 , 1 , - 0.14 , 0 , 0 , 0 , 1 ) " but got "matrix3d ( 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 0 , 1 , - 0.04 , 0 , 0 , 0 , 1 ) "
FAIL Compositing: property <transform> underlying [perspective(10px)] from add [perspective(10px)] to add [perspective(50px)] at (1) should be [perspective(8.33px)] assert_equals: expected "matrix3d ( 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 0 , 1 , - 0.12 , 0 , 0 , 0 , 1 ) " but got "matrix3d ( 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 0 , 1 , - 0.02 , 0 , 0 , 0 , 1 ) "
FAIL Compositing: property <transform> underlying [perspective(10px)] from add [perspective(10px)] to add [perspective(50px)] at (1.5) should be [perspective(10px)] assert_equals: expected "matrix3d ( 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 0 , 1 , - 0.1 , 0 , 0 , 0 , 1 ) " but got "matrix ( 1 , 0 , 0 , 1 , 0 , 0 ) "
FAIL Compositing: property <transform> underlying [perspective(10px)] from accumulate [perspective(10px)] to accumulate [perspective(50px)] at (-0.5) should be [perspective(4.12px)] assert_equals: expected "matrix3d ( 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 0 , 1 , - 0.24 , 0 , 0 , 0 , 1 ) " but got "matrix3d ( 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 0 , 1 , - 0.14 , 0 , 0 , 0 , 1 ) "
FAIL Compositing: property <transform> underlying [perspective(10px)] from accumulate [perspective(10px)] to accumulate [perspective(50px)] at (0) should be [perspective(5px)] assert_equals: expected "matrix3d ( 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 0 , 1 , - 0.2 , 0 , 0 , 0 , 1 ) " but got "matrix3d ( 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 0 , 1 , - 0.1 , 0 , 0 , 0 , 1 ) "
FAIL Compositing: property <transform> underlying [perspective(10px)] from accumulate [perspective(10px)] to accumulate [perspective(50px)] at (0.25) should be [perspective(5.45px)] assert_equals: expected "matrix3d ( 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 0 , 1 , - 0.18 , 0 , 0 , 0 , 1 ) " but got "matrix3d ( 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 0 , 1 , - 0.08 , 0 , 0 , 0 , 1 ) "
FAIL Compositing: property <transform> underlying [perspective(10px)] from accumulate [perspective(10px)] to accumulate [perspective(50px)] at (0.5) should be [perspective(6.15px)] assert_equals: expected "matrix3d ( 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 0 , 1 , - 0.16 , 0 , 0 , 0 , 1 ) " but got "matrix3d ( 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 0 , 1 , - 0.06 , 0 , 0 , 0 , 1 ) "
FAIL Compositing: property <transform> underlying [perspective(10px)] from accumulate [perspective(10px)] to accumulate [perspective(50px)] at (0.75) should be [perspective(7.06px)] assert_equals: expected "matrix3d ( 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 0 , 1 , - 0.14 , 0 , 0 , 0 , 1 ) " but got "matrix3d ( 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 0 , 1 , - 0.04 , 0 , 0 , 0 , 1 ) "
FAIL Compositing: property <transform> underlying [perspective(10px)] from accumulate [perspective(10px)] to accumulate [perspective(50px)] at (1) should be [perspective(8.33px)] assert_equals: expected "matrix3d ( 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 0 , 1 , - 0.12 , 0 , 0 , 0 , 1 ) " but got "matrix3d ( 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 0 , 1 , - 0.02 , 0 , 0 , 0 , 1 ) "
FAIL Compositing: property <transform> underlying [perspective(10px)] from accumulate [perspective(10px)] to accumulate [perspective(50px)] at (1.5) should be [perspective(12.5px)] assert_equals: expected "matrix3d ( 1 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , 0 , 0 , 1 , - 0.08 , 0 , 0 , 0 , 1 ) " but got "matrix ( 1 , 0 , 0 , 1 , 0 , 0 ) "
